* scripts/prokgc.go: parse genome and cds
The prokgc.go program now parses TaxID, genome and cds files as a command line
interface built on non POSIX compliant flags.

It roughly gets the GC contents of a 4.8 Mb long genomes and 4600 coding
sequences in 200ms. It aims to replace the Mokhtaria python script which takes
about 20s for a single genome.