set pipeline to download complete genbank genome
I have decided that it would be easier to simply download all complete genome
sequences from genbank and have it on disk rather than downloading it each time
the script is changed.

The pipeline I have setup is the following:

If the file prokaryotes.txt does not exist, download it from NCBI genome reports
servers and extract complete genome sequence only. Then for each line of this
file, get accession numbers and write them in a file called TAXID.access. Then
use raa_query to query GenBank to download complete genome sequences for each
accession numbers, and coding sequences only.

The script verifies that genome is not already in folder, to avoid running
avoidable queries.

The expected size of the whole data/ folder is between 30 and 74G.